Fix loading of compiled application
** Fix Compiled App Loading
The Behave CLJS app needs to load *after* the WASM module is
loaded. To do this, we add three new global variables:
- `window.onWASMModuleLoaded` (function) - Adds the <script> tag with
  the app.js file
- `window.onWASMModuleLoadedPath` (string) - Path to the app.js file
- `window.onAppLoaded` (function) - Function to run once the app.js
file has been completed.

Then, when the WASM Module is loaded, we set up `onWASMModuleLoaded`
to be called via
[Module.onRuntimeInitalized](https://emscripten.org/docs/api_reference/module.html#Module.onRuntimeInitialized),
the canonical method of running application code once the WASM code
has been set up and initialized.

The process now goes
WASM -> Add Script tag with app.js -> Load app.js -> call client.cljs init method

This process also required a bit of re-jiggering when the solver modules are
loaded. In the future, we may want to cache the results somewhere.

** Fixes Favicon Issue
Also fixes issues with the favicon, adds a new iOS/Android version of
a basic icon (to be replaced) for saving to the home screen.